backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
gitpython>=3.1.31
typing-extensions>=4.5.0 